Resynced with GNU Classpath: small cleanups in File
2004-11-11  Dalibor Topic  <robilad@kaffe.org>

        * libraries/javalib/java/io/File.java:
        Resynced with GNU Classpath.

        2004-11-09  Tom Tromey  <tromey@redhat.com>

        * java/io/File.java (directory.getName): Removed redundant cast.
        Changed mask to be of type long.

+++ libraries/javalib/java/io/File.java
@@ -1118,7 +1118,7 @@ public class File implements Serializable, Comparable
         if (prefix.length() >= 8)
           throw new IllegalArgumentException("Prefix too long: " + prefix + "(valid length 3..7)");
 
-        int  mask = (int) (0x000000ffffFFFFL >> (long) (prefix.length() * 4));
+        long mask = 0x000000ffffFFFFL >> (prefix.length() * 4);
         do
           {
             int n = (int) (System.currentTimeMillis() & mask);
